Refactor folder structure, fixed gitignores and added some tests
