+++ .gitignore
@@ -1,11 +1,28 @@
-dist
+.env
+useful_info+scripts+TODO.md
 multiplepdf
+
+# javaScript
+dist
+
+# python
 tests.ipynb
 gulagcleaner.egg-info
-gulagcleaner/__pycache__
-target
-useful_info+scripts+TODO.md
-.env
+__pycache__
 .venv
-python/gulagcleaner/__pycache__
 *.pyd
+
+#rust
+
+debug/
+target/
+
+# Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
+# More information here https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
+Cargo.lock
+
+# These are backup files generated by rustfmt
+**/*.rs.bk
+
+# MSVC Windows builds of rustc generate these, which store debugging information
+*.pdb
